move: offer - model
