+++ app/dependencies/order_dependency.py
@@ -5,7 +5,8 @@ from dataclasses import dataclass
 from fastapi import Depends, HTTPException, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from models import OrderModel, OfferLinkModel
+from models import OrderModel
+from entities.offer.offer_model import OfferLinkModel
 from models.choices import StatusOrder
 from models.database import get_session
 from models.order_model import OrderInSchema
