+++ app/entities/offer/offer_model.py
