+++ app/models/__init__.py
@@ -1,4 +1,3 @@
-from .offer_model import OfferLinkModel
 from .order_model import OrderModel, OrderDetailModel
 
 # User [Client, Master] (email)
