+++ app/services/stub_init_service.py
@@ -1,10 +1,10 @@
 import datetime as dt
 from decimal import Decimal
 
+from entities.offer.offer_model import OfferLinkModel
 from entities.service_name.models_service_name import ServiceNameModel
 from models import (
     OrderModel,
-    OfferLinkModel,
     OrderDetailModel,
 )
 from entities.category.models_category import CategoryModel
